CB-29457 Fixed env service Regexp Security hotspot. Made long conversions more explicit. Renamed finalize() methods to more proper names for CCM services.

+++ core/src/main/java/com/sequenceiq/cloudbreak/core/flow2/stack/upscale/StackUpscaleService.java
@@ -319,7 +319,7 @@ public class StackUpscaleService {
         int removableCpuCount = additionalRequired + quotaExceededException.getCurrentUsage() - quotaExceededException.getCurrentLimit();
         int removableNodeCount = (int) Math.ceil((double) removableCpuCount / cpuCountPerNode);
 
-        long provisionableNodeCount = originalRequestedNodeCount - removableNodeCount;
+        long provisionableNodeCount = (long) originalRequestedNodeCount - removableNodeCount;
         switch (adjustmentTypeWithThreshold.getAdjustmentType()) {
             case EXACT:
                 if (adjustmentTypeWithThreshold.getThreshold() > provisionableNodeCount) {
